requirements.txt
nose==1.3.7
Click
requests==2.27.1
urllib3==1.26.12
BeautifulReport==0.1.3
marmota@git+https://git.youle.game/opensource/marmota.git@master
PyJWT~=2.4.0
retrying~=1.3.4
